# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Powersport Bike Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BikeSpecs {
  engine?: string;
  horsepower?: string;
  torque?: string;
  weight?: string;
  fuelCapacity?: string;
  seatHeight?: string;
  transmission?: string;
  brakes?: string;
  suspension?: string;
  yearModel?: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  price: number;
  category: string;
  description: string;
  engineCapacity?: string;
  topSpeed?: string;
  images: string[];
  model3dUrl?: string | null;
  featured: boolean;
  specs?: BikeSpecs;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductInput {
  name: string;
  brand: string;
  price: number;
  category: string;
  description: string;
  engineCapacity?: string;
  topSpeed?: string;
  images: string[];
  model3dUrl?: string | null;
  featured?: boolean;
  specs?: BikeSpecs;
}

export interface UpdateProductInput {
  name?: string;
  brand?: string;
  price?: number;
  category?: string;
  description?: string;
  engineCapacity?: string;
  topSpeed?: string;
  images?: string[];
  model3dUrl?: string | null;
  featured?: boolean;
  specs?: BikeSpecs;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  limit: number;
  offset: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface BikeRequest {
  id: number;
  name: string;
  email: string;
  phone: string;
  location: string;
  bikeId?: number | null;
  bikeName?: string | null;
  message?: string | null;
  status: string;
  createdAt: string;
}

export interface CreateRequestInput {
  name: string;
  email: string;
  phone: string;
  location: string;
  bikeId?: number | null;
  bikeName?: string | null;
  message?: string | null;
}

export interface UpdateRequestStatusInput {
  status: string;
}

export interface SocialLinks {
  instagram?: string | null;
  twitter?: string | null;
  website?: string | null;
  facebook?: string | null;
}

export interface AdminProfile {
  id: number;
  name: string;
  email: string;
  phone: string;
  socialLinks?: SocialLinks;
  whatsappNumber?: string;
}

export interface UpdateAdminProfileInput {
  name?: string;
  email?: string;
  phone?: string;
  whatsappNumber?: string;
  socialLinks?: SocialLinks;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  admin: AdminProfile;
  /** Bearer token for cross-origin admin auth (use when session cookies are blocked) */
  token: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DashboardStats {
  totalProducts: number;
  totalRequests: number;
  pendingRequests: number;
  featuredProducts: number;
  popularCategory?: string;
  recentRequestsCount: number;
  requestsByStatus: StatusCount[];
  productsByCategory: CategoryCount[];
}

export interface ActivityDataPoint {
  date: string;
  requests: number;
  products: number;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetProductsParams = {
  category?: string;
  brand?: string;
  minPrice?: number;
  maxPrice?: number;
  search?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type GetRequestsParams = {
  status?: string;
  bikeId?: number;
};

export type GetActivityDataParams = {
  days?: number;
};