# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Powersport Bike Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products for a given product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /admin/profile:
    get:
      operationId: getAdminProfile
      tags: [admin]
      summary: Get admin profile
      responses:
        "200":
          description: Admin profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
    put:
      operationId: updateAdminProfile
      tags: [admin]
      summary: Update admin profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminProfileInput"
      responses:
        "200":
          description: Admin profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current authenticated admin
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: string
        engineCapacity:
          type: string
        topSpeed:
          type: string
        images:
          type: array
          items:
            type: string
        model3dUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        specs:
          $ref: "#/components/schemas/BikeSpecs"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - brand
        - price
        - category
        - description
        - images
        - featured
        - createdAt
        - updatedAt
    BikeSpecs:
      type: object
      properties:
        engine:
          type: string
        horsepower:
          type: string
        torque:
          type: string
        weight:
          type: string
        fuelCapacity:
          type: string
        seatHeight:
          type: string
        transmission:
          type: string
        brakes:
          type: string
        suspension:
          type: string
        yearModel:
          type: string
    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: string
        engineCapacity:
          type: string
        topSpeed:
          type: string
        images:
          type: array
          items:
            type: string
        model3dUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        specs:
          $ref: "#/components/schemas/BikeSpecs"
      required:
        - name
        - brand
        - price
        - category
        - description
        - images
    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: string
        engineCapacity:
          type: string
        topSpeed:
          type: string
        images:
          type: array
          items:
            type: string
        model3dUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        specs:
          $ref: "#/components/schemas/BikeSpecs"
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: number
        limit:
          type: number
        offset:
          type: number
      required:
        - products
        - total
        - limit
        - offset
...
    AdminProfile:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
        whatsappNumber:
          type: string
      required:
        - id
        - name
        - email
        - phone
    SocialLinks:
      type: object
      properties:
        instagram:
          type: string
          nullable: true
        twitter:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        facebook:
          type: string
          nullable: true
    UpdateAdminProfileInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
    AdminLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        admin:
          $ref: "#/components/schemas/AdminProfile"
        token:
          type: string
          description: Bearer token for cross-origin admin auth (use when session cookies are blocked)
      required:
        - success
        - admin
        - token
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message