# lib/api-zod/src/generated/types/adminLoginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Powersport Bike Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminProfile } from "./adminProfile";

export interface AdminLoginResponse {
  success: boolean;
  admin: AdminProfile;
  /** Bearer token for cross-origin admin auth (use when session cookies are blocked) */
  token: string;
}